requirements.txt
potassium
torch
transformers
torchvision
insightface==0.7.3
onnxruntime-gpu
opencv-python
gfpgan